requirements.txt
# Core dependencies
web3>=6.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Smart contract interaction
eth-account>=0.8.0
eth-utils>=2.0.0
web3[tester]>=6.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0

# Code quality
flake8>=6.0.0
black>=23.0.0
mypy>=1.0.0

# Logging and monitoring
structlog>=23.1.0
sentry-sdk>=1.20.0

# Optional: ML for predictions
numpy>=1.24.0
scikit-learn>=1.2.0
